feat: server should reject non-TLS connections by default (#383)
The "use_tls" parameter has been renamed to "require_ssl" and when true
(default) it will configure the server to require an SSL connection. It
can be set to false, in which case the server will accept SSL or
non-SSL connections. If "parameter_group_name" is specified then it will
not take effect.

[#182758757](https://www.pivotaltracker.com/story/show/182758757)