feat: allow s3 encryption value to be removed (#388)
* feat: allow s3 encryption value to be removed

Previously we relied on null values to conditionally
create the encryption resource. CSB does not allow you
to update a value with a null value, so there was no way to
remove encryption once created. By using an empty string as the
default value, users can now remove encryption if required.

[#182296596](https://www.pivotaltracker.com/story/show/182296596)

* update test cases

[#182296596](https://www.pivotaltracker.com/story/show/182296596)