feat: allow s3 encryption value to set to null (#399)
This relates to the properties:
- sse_default_kms_master_key_id
- sse_default_algorithm

These are string values and can be set. When trying to unset them, there
has been some debate as to whether they should be set to "null" or
whether they should be set to the empty string "". Since the
introduction of a new feature allowing strings to be nullable, we have
decided that the best way to unset these properties is to set them to
"null". This change marks the fields as nullable, and mostly reverts PR #388

See: https://github.com/cloudfoundry/cloud-service-broker/pull/584

[#182296596](https://www.pivotaltracker.com/story/show/182296596)