feat: make "require_ssl" property backwards-compatible
- Now defaults to "false" so that anyone upgrading doesn't unexpectedly
end up with apps failing to connect
- Removed the associated SCRAM password - because it's not the default,
and it shouldn't unexpectedly be turned on
- Added a note to the "parameter_group_name" property to explain that it
should not be introduced in an instance update. This is because doing so
will trigger the deletion of the current parameter group, but that will
fail as the current group will be in use until the database is rebooted.

[#182854750](https://www.pivotaltracker.com/story/show/182854750)