feat: expose storage_type (#448)
* feat: expose storage_type

Previously all Postgres instances were created with the
storage type gp2. This is now configurable, as well as exposing the
iops value which is used when the storage_type is 'io1'.
The default for new plans will make use of 'io1'. As a result the
default storage has increased to 100GB as this is the smallest value allowed
for storage_type 'io1'.
* docs: add release notes for storage_type property