feat:Add option to enable audit log export to cloudwatch [#183457241] We want to enable users to be able to export their audit logs for a service instance. This adds the required resources and input variables to enable exporting audit logs to cloudwatch. This setting requires providing a properly configured option_group identified by `option_group_name` (enabling the audit log plugin, (see)[https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/Appendix.MySQL.Options.AuditPlugin.html] )
Known Issues:

IF an instance is created without audit log enabled
AND then the Instance get's updated to enable audit logging
AND then the Instance get's updated to disable audit logging

THEN the log group created in step 2 will NOT be deleted.

This will create an issue if you try to set audit logging to
true again because terraform will error trying to create an
existing log group.

To avoid this, the log groups use name_prefix rather than name.