feat: Aurora MySQL binding to reader endpoint (#550)
Aurora has the concept of Writer nodes and Reader nodes. Reader nodes
are useful for High Availability, but they can also be used for
read-only workloads. This features exposes the ability to bind to the
reader endpoint for read-only workloads. This is achieved by passing the
parameters '{"reader":true}' to the 'cf bind-service' command.

[#183390030](https://www.pivotaltracker.com/story/show/183390030)

Co-authored-by: Felisia Martini <fmartini@pivotal.io>