fix: Removed limitation of redis version
The version was not being used to set the engine and it was relying on the default AWS. The engine was used to create the parameter group name.
Now the version is used correctly to select the engine version and the parameter group name is not set as the default will be correct and based on the engine version.

[#183758518](https://www.pivotaltracker.com/story/show/183758518)