fix: remove failing example after removing plans
[#184405164](https://www.pivotaltracker.com/story/show/184405164)

After removing built-in plans an "example" plan was created and is used
to run every example. But this plan specifies the node_count parameter
so it can't be modified at provisioning level.

I have decided to delete the example to not saturate the marketplace
and CI pipeles with several example plans with different configurations
for the same service.