feat: defaults when node_count is 1 (#859)
When node_count == 1, the multi_az_enabled and
automatic_failover_enabled properties must be false. During testing of
upgrades, we found that this could be problematic for
automatic_failover_enabled.

In previous versions automatic_failover_enabled was false when
node_count was 1 and true otherwise. It was not exposed as a property.

We had a desire to expose automatic_failover_enabled and
multi_az_enabled as properties, and to remove this defaulting behaviour
so that users could be clear that when they set multi_az_enabled or
automatic_failover_enabled it would apply, or there would be a failure.

But considering upgrades, there is a scenario where a plan allows the
user to set the node_count. When upgrading this plan, so instances would
need `automatic_failover_enabled: false` in order to upgrade
successfully and some would not. The upgrade process does not allow
properties to be changed, leading to a unhappy upgrade process.

So we decided that automatic_failover_enabled should be false when the
node_count is 1. For consistency we have done the same for
multi_az_enabled, although this property was not problematic.

[#184869846](https://www.pivotaltracker.com/story/show/184869846)