fix: use meta-argument depends_on to control HTTP requests
AWS disables ACLs by default. It is possible to reactivate the `ACLs` but for this, the bucket ownership must be previously modified to `Bucket owner preferred` or `Object writer`.
Because the HTTP requests performed by Terraform are in a random order, we are obliged to add the `depends_on` meta-argument to control interactions between acls and bucket ownership.

[#185029553](https://www.pivotaltracker.com/story/show/185029553)