fix: [AWS] removed prop Terraform Redis provider
[#185273812]

availability_zones has been deprecated for some time
in latest version of the aws terraform provider this
property has been removed.

It doesn't impact the service in any way. Tests were
failing because we were proactively checking for any
possible changes in all properties (even the ones we
are not using).