chore: increase java apps memory
In a previous commit we reduced memory and disk for all apps.
However java apps need more memory than their go counterparts.
We could investigate and twik the JVM to use more memory but
for now we will just give the apps more memory and see if that
still keeps our pipelines green.

[#185528524](https://www.pivotaltracker.com/story/show/185528524)