fix: [Aurora] instance_name too long corner-case
[#185286271]

https://github.com/cloudfoundry/csb-brokerpak-aws/pull/1042#issuecomment-1660452452

A customer could pass an instance_name of length 63 & `cluster_instances: 2`
(which is theoretically valid), but terraform will fail to apply changes with
an error about cluster's instance_name being too long.

This is because each cluster instance name would be 63 + prefix characters long.
Since the prefix is based on the input variable var.cluster_instances we can't
a fixed max lenght for the cluster identifier with 100% guarantees to never fail.

Instead, we have implemented this corner-case as a precondition.