feat: poc binding functionality
Proof of concept to understand how bindings work in MSSQL.
The provider developed in the Azure Brokerpak has been used as a coding
base with slight modifications for the connection chain.
The database is not created when the servier is created. So it is
necessary to execute the create database sentence.
It is necessary to use an option group called "contained database
authentication" to create the contained database.

[#185149821](https://www.pivotaltracker.com/story/show/185149821)