feat: [MSSQL] dont mask invalid iops combinations
[#185149472]

So far, other services preferred ignoring iops field when storage
type was not io1 or gp3. This implementation raises error instead.

The reasoning is to follow the principle of least surprise. If a
specific combination is invalid, enforce users to clearly specify
correct values.

There is some level of surprise though, because iops is 3000 by
default, so a customer might not be aware of this and still get
the error. However, even in that case, it might be useful to
surface the error to make the default value explicit and evident.