fix: [MSSQL] encrypt instance require db.t3.xlarge
[#185149632]

We used sqlserver-ex with minimum size for testing to minimise costs.
But sqlserver-ex doesn't support encryption which is now the default.

We could disable encryption in our tests and call it a go, but testing
encryption seems a good tradeoff since the costs are not high either,
is the default behaviour and also a common industry security practice.

PR #1059 changed engine to use sqlserver-se but the smallest
instance_class for sqlserver-se is db.t3.xlarge so it failed.