feat: [MSSQL] expose support for storage autoscale
[#185149499]

This implementation doesn't follow conventions stablished in
mysql or postgresql on how to expose storage autoscale.

A larger discussion of the reasons can be found in the tracker
story. The summary is we no longer need to.

We can now define nullable numbers which were impossible before
we can also use Terraform preconditions to validate inputs, etc