fix: [MSSQL] remove unnecessary validations
[#185149472]

Terraform already checks more combinations than us, and do it
fast without having to wait for an error from th IAAS.
Removing these redundant and incomplete validations.

Also allowed iops to be set to null. This is to allow customers
to override the default iops value, in case they want to use
a non-iops based storage_type such as gp2 or standard.