feat: [MSSQL] add containment validation
[#185201223](https://www.pivotaltracker.com/story/show/185201223)

This provider was created assuming that the database and all
users created by it were using containment authentication.

If this condition is not satisfied it is unclear whether some of
the functionality might break, become harder to debug, or behave
in unpredictable ways.

For that reason, this validation was introduced.
It will failfast in the following scenarios (possibly others):
- When using a custom parameter group without containment enabled
- When a non-contained db/service is migrated to this provider
- When a customer accidentally/voluntarily disables containment