chore: use latest version of honnef's staticcheck
Some make targets were failing with "invalid Go version".
This issue was reported and fixed in staticcheck's official repo
several months ago: https://github.com/dominikh/go-tools/issues/1431

This commit also changed the way the module is invoked in Makefile
to not need to include this dependency as a dependency in go.mod.