chore: prototype to run end2end terraform tests
Created in the context of the following story:
[#186276329](https://www.pivotaltracker.com/story/show/186276329)
but decoupled from it afterwards and kept as a code reference.

This commit adds a prototype for a new type of terraform-tests.
This new type of tests allow:
- Creating needed Terraform resources before running the test
- Automatically cleaning up any created Terraform resource
- Performing a whole Terraform cycle "Provision + Bind"
- Defining a `run_test.sh` script to describe your test logic
- Passing "Provision" outputs as "Bind" inputs automatically
- A development feedback loop as close to Terraform as possible
- Specifying inputs for the "Provision" step in a terse manner
- Specifying inputs for the "Bind" step in a terse manner
- Parallel execution of the tests (not harnessed by prototype)

This commit adds tests to reproduce the bug we are trying to solve
but doesn't include the fix yet. This commit won't pass the tests