test: only migrate data schema
The database migration test started to fail because it could not copy
across the "performance_insights" schema. This was likely triggered by a
change in AWS, but the underlying cause is that the migration test tries
to copy across all schemas, and we should restrict it to just copying
across the data schema.

[#186350212](https://www.pivotaltracker.com/story/show/186350212)