chore: prevent redis bindings being turned into modules (#1248)
* chore: prevent redis bindings being turned into modules

When specifying `template` or `template_ref` in a service offering yaml,
the broker wraps the terraform file/content into a module. This is
not necessary, and it causes issues when trying to execute provider
renames as terraform would require an init to be performed (and fail)
to install the module (this change was detected in versions > 1.2.0 of
terraform).
Redis bindings do not need any terraform, as the broker already merges
all the provision outputs with the binding outpus and hence hostname,
ports etc would be provided without any additional work. However in the
past we have leaned towards being explicit in our brokerpak definitions
about the outputs of each service, to avoid confusion. So the approach
in this PR is to just add everything that would be expected in a binding
as an output explicitely in the terraform files, and specify them with
`template_refs`, which won't cause the broker to wrap files into a
module.

[#186230750](https://www.pivotaltracker.com/story/show/186230750)

* chore: enable provider renames in AWS tests

[#183286656](https://www.pivotaltracker.com/story/show/183286656)

---------

Co-authored-by: Andrea Zucchini <zandrea@vmware.com>