chore: replace dummy vars and output files with noop files (#1316)
As part of a previous commit we had introduced explicit variables and
output files in the redis terraform templates. This was to have actual
terraform to add to the workspace and prevent the broker to wrap the
code into a module. We since changed the approach to add a noop file
with a comment, as the variables/output appraoch was too verbose and we
were reluctant to do that in all services, risking consistency across
services.

[#186527733](https://www.pivotaltracker.com/story/show/186527733)