Rename `types` to `properties` and hide `inputs` variable
The `inputs` variable is a totally generic placeholder and
doesn't need to be modified across services or iaases.

We can't get rid of it in favour of the `properties` variable.
This is so because specifying the names and types of each field
"tells" Terraform to automatically generate a default value for
each of the fields (even if it is simply a `null` value).

This in turn makes it impossible to distinguish unspecified
properties from properties explicitly passing a null value.

In contrast, the `inputs` variable being of type `any` doesn't
"tell" Terraform anything about its content, so it will contain
exactly the inputs we specify. Nothing more, nothing less.

To prevent users from passing unsupported, arbitrary key-values
we compare both variables: `inputs` and `properties` and raise
an error if `inputs` contains any key not found in `properties`
(we treat `properties` variable as a kind-of type declaration)