chore: make test brokerpak upgrade green (#1483)
* Revert "fix: test trying to downgrade (#1458)"

This reverts commit 41363092a7a1b6d03a4ec8ddf155fddb910d6d8e.

* Revert "chore: workaround issues during aurorapg major upgrade (#1429)"

This reverts commit ece1cd7a03f3b7ed34ef2cd702c968afdece92b6.

* chore: make test-brokerpak-upgrade green

After several days investigating the issue I was not able to
make the major upgrade work unattended.
The following tutorial explains the issues we faced:
https://faun.pub/how-did-i-upgrade-aurora-postgres-rds-clusters-in-production-using-terraform-ff57aeb16873

I found no way to follow the same steps as the tutorial since
one of its steps requires:
- modifying the cluster parameter group to use the default one
- leaving the existing parameter group without deleting it

The current implementation of the service, when you specify a
different parameter group the old one is automatically marked
for deletion.

I haven't found any acceptable workaround to satisfy those
requirements.

I'm modifying the test to elude the issue. Although it is just
a temporary workaround there are two good reasons to accept it:
- unblocks the pipeline to be able to release a new version
- the issue might also be present in aurora-mysql but since
  we're not bumping a major version in its test it has been
  "hidden" all this time.

This issue dats back to April 2023, and the test was created
precisely to test this scenario:
https://github.com/cloudfoundry/csb-brokerpak-aws/pull/886/files

In any case, we should come back and fix this asap to prevent
any customer incidents related to this challenge.