chore(tests): simplify aurora-postgres update and upgrade test
changed to increase capacity on serverless instance instead of
doing extra updates to change instance class. aurora seems to be
particularly problematic by itself on upgrades, even through the
console.
In this test we want to test that the brokerpak has been changed in
a way that allows for updates and upgrades and not aws behavior.

[#186886707](https://www.pivotaltracker.com/story/show/186886707)