chore: use default tags for SQS
By using default tags configured on the provider, we avoid having to add
them to every resource. At this stage it doesn't make a lot of
difference, but if we were to add more resources then it would reduce
duplication.

Note that this cannot be done for bindings (yet) because labels are not
supplied as part of the binding request.