fix: aws terraform-tests performance_insights_retention_period
[#187103734](https://www.pivotaltracker.com/story/show/187103734)

Recently, a new PR for validating this property was merged:
https://github.com/hashicorp/terraform-provider-aws/pull/35870

I believe it's valuable to have our tests break when changes
like this are introduced in the providers. Therefore, we should
try have behaviours such as this encoded in our tests. Things
like: "invalid values for property X are not validated by TF".