fix: allow null values for iops (#1605)
When creating rds instances an iops value of 0 is interpreted by
downstream services as null value. However when updating this is not the
case. Hence when trying to update a service where setting iops is not
permitted by AWS, update operations error even when we are not trying to
change the iops value. Allowing nulls for this property works as
expected both for provisioning and update.

[#187149548](https://www.pivotaltracker.com/story/show/187149548)