Update queue users permissions
The same permissions are given to a Standard/FIFO queue user and to a DLQ user. We have no reasonn at the moment to think that the permissions should be different.
If the instance has a DLQ associated the user will have permissions to trigger redrive from DLQ back to the original queue