feat: SQS allow sqs_managed_sse_enabled to be disabled
Allows the "sqs_managed_sse_enabled" to be disabled on a queue, while
retaining the default of being enabled.

When "kms_master_key_id" is set, it will have the effect of unsetting
"sqs_managed_sse_enabled", rather than failing with a conflict. This is
to make the experience of setting "kms_master_key_id" more
user-friendly.

[#187061621](https://www.pivotaltracker.com/story/show/187061621)