chore: Fixed failing AWS SQS terraform test.
AWS has changed behaviour to return zero for kms_data_key_reuse_period_seconds unless kms_master_key_id is set.
Because of the limitation of how we wire up our brokerpaks with terraform vars, this parameter has to be set and set to a valid value - but AWS willnow return 0 for unless kms_master_key_id is set.

This change corrects the test to match current behaviour.