chore: spread out Dependabot updates
We see a lot of conflicts with Dependabot updates, presumably because
many PRs are created at the same time. Prior to https://github.com/cloudfoundry/csb-brokerpak-aws/pull/1947
we created the PRs at different times, and it feels like we got
fewer conflicts. Therefore re-adding the times for different modules
to ensure that the PRs have time to be merged before the next PR
is created.